+++ example/echo/broadcast.go
@@ -17,7 +17,7 @@ func main() {
 	flag.StringVar(&addr, "addr", ":10010", "echo server address")
 	flag.Parse()
 
-	server, err := link.Serve("tcp", addr, link.Async(codec.String(codec.Uint16BE), 1024))
+	server, err := link.Serve("tcp", addr, link.Async(1024, codec.String(codec.Uint16BE)))
 	if err != nil {
 		panic(err)
 	}